fix(workspace): 修复实体列类型不匹配导致的两处接口400
全面验收发现：1) mxx_system_workspace 表 is_default/status/deleted 为 SMALLINT 而实体声明 Option<i32>，导致 GET /workspace/list 400；2) mxx_system_dashboard_user_layout 表 hidden 为 SMALLINT，导致保存个人布局覆盖后 GET /dashboard/user/layout 400。实体与 model 统一改为 Option<i16> 并补显式转换。修复后全面验收 25/25 通过（M1 布局持久化/隔离/恢复默认、M3 工作台过滤、M4 homePath、权限矩阵、停用卡三端同步缓存失效链路）。

+++ backend/src/modules/system/entity/dashboard_user_layout.rs
@@ -33,7 +33,7 @@ pub struct Model {
     /// 高（行数）
     pub h: Option<i32>,
     /// 是否隐藏（1隐藏 0显示）
-    pub hidden: Option<i32>,
+    pub hidden: Option<i16>,
     /// 更新时间
     pub update_time: Option<DateTime>,
 }
